gri-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20

# --- Database (Python 3.12.7 호환) ---
sqlalchemy[asyncio]==2.0.36 # ✅ [수정] asyncio extras 포함
asyncpg==0.29.0 # ✅ [수정] 동기용 psycopg2-binary 대신 비동기용 asyncpg 사용

# --- AI Model & GPU Inference (Llama3, RTX 5060, CUDA 12.8 호환) ---
torch>=2.5.0
transformers==4.46.3
accelerate==1.2.1
bitsandbytes==0.45.0
huggingface-hub==0.26.5
peft==0.15.2
datasets==3.2.0

# --- Elasticsearch & Data Handling (Python 3.12.7 호환) ---
elasticsearch==8.17.0
pandas==2.2.3
openpyxl==3.1.5

# --- Tokenizer & Utilities (Python 3.12.7 및 CUDA 12.8 호환) ---
sentencepiece==0.2.0
protobuf==5.29.2
numpy==2.1.3

langchain
langchain-community